move test_api.py to helpers/base.py
